Enlarge collocation region due to ice drift speed
collocate.py:
* Added an extra argument, 'max_drift_speed'. This defines the maximum
speed that one allow ice to move [in meters per second].
This is used to enlarge the region of the second image that is cropped
when creating the pairs, in order to allow a larger region to register
with the first image.

RasterAdjuster.py:
* Added the parameter 'intersection_extension' describing how much the
intersection region should be inflated/enlarged.