Allow to set minimum and maximum time separation for images to download.
asf_meta_download.py:
* Allow to download from both s1a and s1b.
* Allow to define the minimum- and maximum- time separation between a
downloaded image and the target image. This is now changed from looking
in an interval of 24 hours before and after, as it was implemented
before.

gtiff_s1_download.py:
* Introduced the arguments for minimum and maximum number of hours time
delay. These values can also be negative if one wants to look prior to
the target image.