Added 'cull_pairs.py' script
Added 'cull_pairs.py' to remove all unecessary pairs in each time slot.
The script is given the time duration of a time slot. Then, it computes
the overlap between the pairs of images. It then removes all image pairs
in each time slot that do not happen to be the maximum overlapping area
for that time slot. In this way, the output will be that at most one
image pair for each time slot will exist.